luet-build: Rewrite `tree pkglist` command
`luet-build tree pkglist` now uses indexes
on read packages list.

For now the options --revdeps, --deps, --buildtime,
--full are no more availables.
